+++ CHANGELOG.md
@@ -15,12 +15,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Triggers: Tools that can be used to shape how the chatbot acts
     - Parrot
     - Stage Directions
-- Added OpenPhonemizer to the download manifest and 3rd party license sheet
 
 #### Changed
 
 - Increased efficiency of LLM with large system prompts
 - TTS can now pronounce words that aren't even in its dictionary because of a new failsafe in the graphene-to-phoneme pipeline.
+    - This causes text with unknown words to take much longer to tokenize, so tokenization order is now strictly enforced in KokoroHandler.GenerateSpeech.
+- Added OpenPhonemizer to the download manifest and 3rd party license sheet
 
 -----