+++ CHANGELOG.md
@@ -6,6 +6,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 -----
 
+## [0.11.0] - ????-??-??
+
+### [TITLE].
+
+#### Added
+
+- Triggers: Tools that can be used to shape how the chatbot acts
+    - Parrot
+    - Stage Directions
+
+#### Changed
+
+- Increased efficiency of LLM with large system prompts
+- TTS can now pronounce words that aren't even in its dictionary because of a new failsafe in the graphene-to-phoneme pipeline.
+    - This causes text with unknown words to take much longer to tokenize, so tokenization order is now strictly enforced in KokoroHandler.GenerateSpeech.
+- Added OpenPhonemizer to the download manifest and 3rd party license sheet
+
+-----
+
 ## [0.10.0] - 2026-06-07
 
 ### Adding the resource downloader.